[IFRT] Define utilities for `absl::Status` manipulation to use IFRT user contexts
This change introduces utilities for manipulating `absl::Status` to attach user
contexts to it.

Based on a design discussion wiht junwhan@.

* `AttachUserContextId`: Used when the (typically low-level) runtime knows a
user context ID to attach to an error.

* `AttachUserContextRef`: Used by the IFRT implementation to attach a strong
reference of `UserContextRef` to `absl::Status` so that it can be expendend
with `ExpandUserContexts` safely by the user.

* `ReattachUserContextRefs`: Used by the IFRT implementation to ensure that the
given `absl::Status` to have strong references to any user contexts previously
attached using `AttachUserContextId` (or by serializing/deserializing
`absl::Status`).

* `ExpandUserContexts`: Used by IFRT users to make `absl::Status` to have the
expanded form of user contexts attached to it. Called when it is safe to do so,
e.g., while holding Python GIL, for `jax::PyUserContext` that works best when
called with GIL already held.

PiperOrigin-RevId: 803612712