Switch to the Tfrt GPU client.
This change also adds a way to detect if the GPU SE client is being used to
conditionally skip specific test cases, and adds a variant `_notfrt` that that
runs GPU tests like before to provide test coverage to the SE GPU client.

In short: this change makes the TFRT GPU client the default for GPU tests that
have been migrated to the PjRt runtime, creating an alternate `_notfrt` target
to continue running tests that do work with the SE GPU PjRt client as before.

PiperOrigin-RevId: 803624437